Move SiteLogo to own component
