+++ src/components/PageWrapper/PageWrapper.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import {Footer} from './Footer'
 import {Header} from './Header'
 import {HeroTitle} from './HeroTitle'
-import {SiteLogo} from './SiteLogo'
 import styles from './PageWrapper.module.scss'
+import {SiteLogo} from '@components/SiteLogo'
 import {ColorSchemeContext} from '@contexts/ColorScheme'
 import {useSafeContext} from '@hooks/use-safe-context'
 import {combineClassNames} from '@utilities/combine-class-names'
